Refactor Dockerfiles, TypeScript config, and service management (#16)
* refactor: update Dockerfiles and TypeScript configuration

- Refactored tsconfig.json to include stricter type checks and additional compiler options.
- Removed tsconfig.base.json as it was no longer needed.
- Updated Docker configurations to streamline development and production builds.
- Move jwt.ts file into src directory.
- Updated frontend dependencies, including upgrading Vite to version 8.0.8 and other related packages.

* refactor: update Makefile and Docker Compose configurations for improved service management
