+++ infra/nginx/conf.d/default.conf
@@ -4,11 +4,6 @@ map $http_upgrade $connection_upgrade {
   ''      close;
 }
 
-upstream api_gateway_upstream {
-  server api-gateway:4000;
-  keepalive 64;
-}
-
 upstream frontend_upstream {
   server frontend:5173;
   keepalive 64;
@@ -75,37 +70,43 @@ server {
     proxy_cache off;
   }
 
-  # -----------------------------
-  # Assets estáticos (cache agresiva)
-  # -----------------------------
-  location ~* \.(js|mjs|css|png|jpg|jpeg|gif|ico|svg|woff|woff2|ttf)$ {
-    expires 7d;
-    add_header Cache-Control "public, max-age=604800, immutable";
-    proxy_pass http://frontend_upstream;
-    proxy_http_version 1.1;
+  # # -----------------------------
+  # # Assets estáticos (cache agresiva)
+  # # -----------------------------
+  # location ~* \.(js|mjs|css|png|jpg|jpeg|gif|ico|svg|woff|woff2|ttf)$ {
+  #   expires 7d;
+  #   add_header Cache-Control "public, max-age=604800, immutable";
+  #   proxy_pass http://frontend_upstream;
+  #   proxy_http_version 1.1;
 
-    proxy_set_header Host $host;
-    proxy_set_header X-Real-IP $remote_addr;
-    proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-    proxy_set_header X-Forwarded-Proto $scheme;
+  #   proxy_set_header Host $host;
+  #   proxy_set_header X-Real-IP $remote_addr;
+  #   proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+  #   proxy_set_header X-Forwarded-Proto $scheme;
 
-    proxy_set_header Upgrade $http_upgrade;
-    proxy_set_header Connection $connection_upgrade;
+  #   proxy_set_header Upgrade $http_upgrade;
+  #   proxy_set_header Connection $connection_upgrade;
 
-    proxy_connect_timeout 5s;
-    proxy_send_timeout 30s;
-    proxy_read_timeout 30s;
-  }
+  #   proxy_connect_timeout 5s;
+  #   proxy_send_timeout 30s;
+  #   proxy_read_timeout 30s;
+  # }
 
-  # -----------------------------
-  # React SPA fallback
-  # -----------------------------
+  # # -----------------------------
+  # # React SPA fallback
+  # # -----------------------------
   # location / {
   #   try_files $uri /index.html;
   # }
 
+  # location ~* \.(js|mjs|css|png|jpg|jpeg|gif|ico|svg|woff|woff2|ttf)$ {
+  #   expires 7d;
+  #   add_header Cache-Control "public, max-age=604800, immutable";
+  #   try_files $uri =404;
+  # }
+
   # -----------------------------
-  # Nginx Test
+  # Development
   # -----------------------------
   location / {
     proxy_pass http://frontend_upstream;
