+++ srcs/backend/services/api-gateway/Dockerfile
@@ -1,56 +1,25 @@
-# FROM node:24-alpine3.22
-
-# WORKDIR /app
-
-# # Dependencias
-# COPY package*.json ./
-# RUN npm install
-
-# # Código
-# COPY . .
-# RUN npm run build
-
-# # Puerto interno del servicio (ajústalo por servicio)
-# EXPOSE 4000
-
-# # Comando dev (requiere script "dev" en package.json)
-# CMD ["npm", "dev"]
-
-# Stage 1: Compilación (Build)
 FROM node:24-alpine3.22 AS builder
-
 WORKDIR /app
-
 COPY package*.json ./
-RUN npm install
-
-COPY . .
+RUN npm ci
+COPY --chown=node:node . .
+USER node
 RUN npm run build
 
-# Stage 2: Producción
 FROM node:24-alpine3.22 AS production
-
 WORKDIR /app
-
 COPY package*.json ./
-RUN npm install --only=production
-
+RUN npm ci --omit=dev
 COPY --from=builder /app/dist ./dist
-
+USER node
 EXPOSE 4000
-
 CMD ["node", "dist/index.js"]
 
-# Stage 3: Desarrollo
 FROM node:24-alpine3.22 AS development
-
 WORKDIR /app
-
 COPY package*.json ./
-RUN npm install
-
-COPY . .
-
+RUN npm ci
+COPY --chown=node:node . .
+USER node
 EXPOSE 4000
-
 CMD ["npm", "run", "dev"]
