+++ srcs/backend/services/auth-service/.env.example
@@ -1,0 +1,1 @@
+JWT_SECRET=jwt_secret
