+++ srcs/backend/services/auth-service/src/lib/jwt.ts
