+++ srcs/frontend/package.json
@@ -25,6 +25,6 @@
     "globals": "^17.4.0",
     "typescript": "~5.9.3",
     "typescript-eslint": "^8.57.0",
-    "vite": "^8.0.1"
+    "vite": "^8.0.8"
   }
 }
